refactor: split up mob.go
