+++ branch.go
@@ -1,0 +1,249 @@
+package main
+
+import (
+	"fmt"
+	"strconv"
+	"strings"
+)
+
+type Branch struct {
+	Name string
+}
+
+func newBranch(name string) Branch {
+	return Branch{
+		Name: strings.TrimSpace(name),
+	}
+}
+
+func (branch Branch) String() string {
+	return branch.Name
+}
+
+func (branch Branch) Is(branchName string) bool {
+	return branch.Name == branchName
+}
+
+func (branch Branch) remote(configuration Configuration) Branch {
+	return newBranch(configuration.RemoteName + "/" + branch.Name)
+}
+
+func (branch Branch) hasRemoteBranch(configuration Configuration) bool {
+	remoteBranches := gitRemoteBranches()
+	remoteBranch := branch.remote(configuration).Name
+	debugInfo("Remote Branches: " + strings.Join(remoteBranches, "\n"))
+	debugInfo("Remote Branch: " + remoteBranch)
+
+	for i := 0; i < len(remoteBranches); i++ {
+		if remoteBranches[i] == remoteBranch {
+			return true
+		}
+	}
+
+	return false
+}
+
+func (branch Branch) IsWipBranch(configuration Configuration) bool {
+	if branch.Name == "mob-session" {
+		return true
+	}
+
+	return strings.Index(branch.Name, configuration.WipBranchPrefix) == 0
+}
+
+func (branch Branch) addWipPrefix(configuration Configuration) Branch {
+	return newBranch(configuration.WipBranchPrefix + branch.Name)
+}
+
+func (branch Branch) addWipQualifier(configuration Configuration) Branch {
+	if configuration.customWipBranchQualifierConfigured() {
+		return newBranch(addSuffix(branch.Name, configuration.wipBranchQualifierSuffix()))
+	}
+	return branch
+}
+
+func addSuffix(branch string, suffix string) string {
+	return branch + suffix
+}
+
+func (branch Branch) removeWipPrefix(configuration Configuration) Branch {
+	return newBranch(removePrefix(branch.Name, configuration.WipBranchPrefix))
+}
+
+func removePrefix(branch string, prefix string) string {
+	if !strings.HasPrefix(branch, prefix) {
+		return branch
+	}
+	return branch[len(prefix):]
+}
+
+func (branch Branch) removeWipQualifier(localBranches []string, configuration Configuration) Branch {
+	for !branch.exists(localBranches) && branch.hasWipBranchQualifierSeparator(configuration) {
+		afterRemoval := branch.removeWipQualifierSuffixOrSeparator(configuration)
+
+		if branch == afterRemoval { // avoids infinite loop
+			break
+		}
+
+		branch = afterRemoval
+	}
+	return branch
+}
+
+func (branch Branch) removeWipQualifierSuffixOrSeparator(configuration Configuration) Branch {
+	if !configuration.customWipBranchQualifierConfigured() { // WipBranchQualifier not configured
+		return branch.removeFromSeparator(configuration.WipBranchQualifierSeparator)
+	} else { // WipBranchQualifier not configured
+		return branch.removeWipQualifierSuffix(configuration)
+	}
+}
+
+func (branch Branch) removeFromSeparator(separator string) Branch {
+	return newBranch(branch.Name[:strings.LastIndex(branch.Name, separator)])
+}
+
+func (branch Branch) removeWipQualifierSuffix(configuration Configuration) Branch {
+	if strings.HasSuffix(branch.Name, configuration.wipBranchQualifierSuffix()) {
+		return newBranch(branch.Name[:strings.LastIndex(branch.Name, configuration.wipBranchQualifierSuffix())])
+	}
+	return branch
+}
+
+func (branch Branch) exists(existingBranches []string) bool {
+	return stringContains(existingBranches, branch.Name)
+}
+
+func (branch Branch) hasWipBranchQualifierSeparator(configuration Configuration) bool { //TODO improve (dont use strings.Contains, add tests)
+	return strings.Contains(branch.Name, configuration.WipBranchQualifierSeparator)
+}
+
+func (branch Branch) hasLocalCommits(configuration Configuration) bool {
+	local := silentgit("for-each-ref", "--format=%(objectname)", "refs/heads/"+branch.Name)
+	remote := silentgit("for-each-ref", "--format=%(objectname)", "refs/remotes/"+branch.remote(configuration).Name)
+	return local != remote
+}
+
+func (branch Branch) hasUnpushedCommits(configuration Configuration) bool {
+	countOutput := silentgit(
+		"rev-list", "--count", "--left-only",
+		"refs/heads/"+branch.Name+"..."+"refs/remotes/"+branch.remote(configuration).Name,
+	)
+	unpushedCount, err := strconv.Atoi(countOutput)
+	if err != nil {
+		panic(err)
+	}
+	unpushedCommits := unpushedCount != 0
+	if unpushedCommits {
+		sayInfo(fmt.Sprintf("there are %d unpushed commits on local base branch <%s>", unpushedCount, branch.Name))
+	}
+	return unpushedCommits
+}
+
+func stringContains(list []string, element string) bool {
+	found := false
+	for i := 0; i < len(list); i++ {
+		if list[i] == element {
+			found = true
+		}
+	}
+	return found
+}
+
+func (branch Branch) isOrphanWipBranch(configuration Configuration) bool {
+	return branch.IsWipBranch(configuration) && !branch.hasRemoteBranch(configuration)
+}
+
+func branch(configuration Configuration) {
+	say(silentgit("branch", "--list", "--remote", newBranch("*").addWipPrefix(configuration).remote(configuration).Name))
+
+	// DEPRECATED
+	say(silentgit("branch", "--list", "--remote", newBranch("mob-session").remote(configuration).Name))
+}
+
+func determineBranches(currentBranch Branch, localBranches []string, configuration Configuration) (baseBranch Branch, wipBranch Branch) {
+	if currentBranch.Is("mob-session") || (currentBranch.Is("master") && !configuration.customWipBranchQualifierConfigured()) {
+		// DEPRECATED
+		baseBranch = newBranch("master")
+		wipBranch = newBranch("mob-session")
+	} else if currentBranch.IsWipBranch(configuration) {
+		baseBranch = currentBranch.removeWipPrefix(configuration).removeWipQualifier(localBranches, configuration)
+		wipBranch = currentBranch
+	} else {
+		baseBranch = currentBranch
+		wipBranch = currentBranch.addWipPrefix(configuration).addWipQualifier(configuration)
+	}
+
+	debugInfo("on currentBranch " + currentBranch.String() + " => BASE " + baseBranch.String() + " WIP " + wipBranch.String() + " with allLocalBranches " + strings.Join(localBranches, ","))
+	if currentBranch != baseBranch && currentBranch != wipBranch {
+		// this is unreachable code, but we keep it as a backup
+		panic("assertion failed! neither on base nor on wip branch")
+	}
+	return
+}
+
+func getWipBranchesForBaseBranch(currentBaseBranch Branch, configuration Configuration) []string {
+	remoteBranches := gitRemoteBranches()
+	debugInfo("check on current base branch " + currentBaseBranch.String() + " with remote branches " + strings.Join(remoteBranches, ","))
+
+	remoteBranchWithQualifier := currentBaseBranch.addWipPrefix(configuration).addWipQualifier(configuration).remote(configuration).Name
+	remoteBranchNoQualifier := currentBaseBranch.addWipPrefix(configuration).remote(configuration).Name
+	if currentBaseBranch.Is("master") {
+		// LEGACY
+		remoteBranchNoQualifier = "mob-session"
+	}
+
+	var result []string
+	for _, remoteBranch := range remoteBranches {
+		if strings.Contains(remoteBranch, remoteBranchWithQualifier) || strings.Contains(remoteBranch, remoteBranchNoQualifier) {
+			result = append(result, remoteBranch)
+		}
+	}
+
+	return result
+}
+
+func hasLocalBranch(localBranch string) bool {
+	localBranches := gitBranches()
+	debugInfo("Local Branches: " + strings.Join(localBranches, "\n"))
+	debugInfo("Local Branch: " + localBranch)
+
+	for i := 0; i < len(localBranches); i++ {
+		if localBranches[i] == localBranch {
+			return true
+		}
+	}
+
+	return false
+}
+
+func gitBranches() []string {
+	return strings.Split(silentgit("branch", "--format=%(refname:short)"), "\n")
+}
+
+func gitRemoteBranches() []string {
+	return strings.Split(silentgit("branch", "--remotes", "--format=%(refname:short)"), "\n")
+}
+
+func gitCurrentBranch() Branch {
+	// upgrade to branch --show-current when git v2.21 is more widely spread
+	return newBranch(silentgit("rev-parse", "--abbrev-ref", "HEAD"))
+}
+
+func gitCommitHash() string {
+	return silentgitignorefailure("rev-parse", "HEAD")
+}
+
+func warnForActiveWipBranches(configuration Configuration, currentBaseBranch Branch) {
+	if isMobProgramming(configuration) {
+		return
+	}
+
+	// TODO show all active wip branches, even non-qualified ones
+	existingWipBranches := getWipBranchesForBaseBranch(currentBaseBranch, configuration)
+	if len(existingWipBranches) > 0 && configuration.WipBranchQualifier == "" {
+		sayWarning("Creating a new wip branch even though preexisting wip branches have been detected.")
+		for _, wipBranch := range existingWipBranches {
+			sayWithPrefix(wipBranch, "  - ")
+		}
+	}
+}
